Moved all lifecycle services to backend directory
