+++ main.py
@@ -1,4 +1,4 @@
-from src.lifecycle import ApplicationLifecycle
+from src.backend.lifecycle import ApplicationLifecycle
 
 if __name__ == "__main__":
     app = ApplicationLifecycle(period=2)
