+++ requirements.txt
@@ -23,3 +23,5 @@ urllib3==2.3.0
 virtualenv==20.29.1
 shapely==2.1.1
 SQLAlchemy==2.0.37
+dash==3.2.0
+plotly==6.5.0
