+++ src/backend/clients/service_client.py
