+++ src/backend/clients/windows_webcam_client.py
@@ -2,7 +2,7 @@ import cv2
 import numpy as np
 import requests
 
-from src.clients.service_client import ServiceClient
+from src.backend.clients.service_client import ServiceClient
 
 
 class WindowsWebcamClient(ServiceClient):
