+++ src/backend/user_interfaces/profile_creation.py
@@ -3,9 +3,9 @@ from typing import List, Tuple
 
 from screeninfo import Monitor
 
-from src.clients.vision_tracking_client import VisionTrackingClient
-from src.clients.windows_webcam_client import WindowsWebcamClient
-from src.user_interfaces.gui_base_class import BaseGUITest
+from src.backend.clients.vision_tracking_client import VisionTrackingClient
+from src.backend.clients.windows_webcam_client import WindowsWebcamClient
+from src.backend.user_interfaces.gui_base_class import BaseGUITest
 
 
 class ProfileCreationGUI(BaseGUITest):
