+++ tests/test_lifecycle.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import MagicMock, patch
 
-from src.lifecycle import ApplicationLifecycle
+from src.backend.lifecycle import ApplicationLifecycle
 
 
 class TestApplicationLifecycleHealthCheck(unittest.TestCase):
@@ -11,21 +11,21 @@ class TestApplicationLifecycleHealthCheck(unittest.TestCase):
         # Patching to run in headless environments.
         dummy_monitor = MagicMock()
         with patch(
-            "src.screen_region.MonitorUtility.select_monitor",
+            "src.backend.screen_region.MonitorUtility.select_monitor",
             return_value=dummy_monitor,
         ):
             self.app = ApplicationLifecycle(period=0.1)
 
     @patch(
-        "src.clients.vision_tracking_client.VisionTrackingClient.get_service_status",
+        "src.backend.clients.vision_tracking_client.VisionTrackingClient.get_service_status",
         return_value=True,
     )
     @patch(
-        "src.clients.windows_webcam_client.WindowsWebcamClient.get_service_status",
+        "src.backend.clients.windows_webcam_client.WindowsWebcamClient.get_service_status",
         return_value=True,
     )
     @patch(
-        "src.clients.system_watchdog_client.SystemWatchdogClient.get_service_status",
+        "src.backend.clients.system_watchdog_client.SystemWatchdogClient.get_service_status",
         return_value=True,
     )
     def test_positive_global_health(
@@ -38,15 +38,15 @@ class TestApplicationLifecycleHealthCheck(unittest.TestCase):
         watchdog_get.assert_called_once()
 
     @patch(
-        "src.clients.vision_tracking_client.VisionTrackingClient.get_service_status",
+        "src.backend.clients.vision_tracking_client.VisionTrackingClient.get_service_status",
         return_value=False,
     )
     @patch(
-        "src.clients.windows_webcam_client.WindowsWebcamClient.get_service_status",
+        "src.backend.clients.windows_webcam_client.WindowsWebcamClient.get_service_status",
         return_value=False,
     )
     @patch(
-        "src.clients.system_watchdog_client.SystemWatchdogClient.get_service_status",
+        "src.backend.clients.system_watchdog_client.SystemWatchdogClient.get_service_status",
         return_value=False,
     )
     def test_negative_global_health(
